requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
numpy>=1.25.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
tqdm>=4.65.0